chore: refactor code and deal with circular dependency
